Move all classes to a sub-namespace
