+++ docs/implement-import.md
@@ -1,4 +1,4 @@
 ## Implement import
 
-* Create a new class `lib/service/BoardImport<ImportName>Service.php` where `<ImportName>` is the name of the source system.
-* Use the `lib/service/BoardImportTrelloJsonService.php` class as inspiration
+* Create a new class `lib/Service/Importer/Systems/<ImportName>Service.php` where `<ImportName>` is the name of the source system.
+* Use the `lib/Service/Importer/Systems/TrelloJsonService.php` class as inspiration
