+++ docs/resources/BoardImport.yuml
@@ -13,12 +13,12 @@
 [BoardImport]uses-.->[BoardImportCommandService|+bootstrap();+import();+validateSystem();#validateConfig();#validateData()]
 [BoardImportCommandService]->[BoardImportService]
 
-[BoardImportService]uses-.->[BoardImportTrelloApiService|+name:string]
-[BoardImportTrelloApiService]uses-.->[BoardImportService]
-[BoardImportTrelloApiService]implements-.-^[<<abstract>> ABoardImportService|#needValidateData:false|+needValidateData():bool]
+[BoardImportService]uses-.->[TrelloApiService|+name:string]
+[TrelloApiService]uses-.->[BoardImportService]
+[TrelloApiService]implements-.-^[<<abstract>> ABoardImportService|#needValidateData:false|+needValidateData():bool]
 
-[BoardImportService]uses-.->[BoardImportTrelloJsonService|+name:string;#needValidateData:true]
-[BoardImportTrelloJsonService]uses-.->[BoardImportService]
+[BoardImportService]uses-.->[TrelloJsonService|+name:string;#needValidateData:true]
+[TrelloJsonService]uses-.->[BoardImportService]
 [BoardImportService]-[note: validateSystem is public because is used on Api. {bg:cornsilk}]
-[BoardImportTrelloJsonService]-[note: To create an import to another system, create another class similar to this. {bg:cornsilk}]
-[BoardImportTrelloJsonService]implements-.-^[<<abstract>> ABoardImportService]
+[TrelloJsonService]-[note: To create an import to another system, create another class similar to this. {bg:cornsilk}]
+[TrelloJsonService]implements-.-^[<<abstract>> ABoardImportService]
