+++ tests/unit/controller/BoardImportApiControllerTest.php
@@ -1,23 +1,23 @@
 <?php
 /**
- * @copyright Copyright (c) 2018 Ryan Fletcher <ryan.fletcher@codepassion.ca>
+ * @copyright Copyright (c) 2021 Vitor Mattos <vitor@php.rio>
  *
- * @author Ryan Fletcher <ryan.fletcher@codepassion.ca>
+ * @author Vitor Mattos <vitor@php.rio>
  *
  * @license GNU AGPL version 3 or any later version
  *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or (at your option) any later version.
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
  *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
 namespace OCA\Deck\Controller;
@@ -26,14 +26,14 @@ use OCA\Deck\Db\Board;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\DataResponse;
 use OCP\IRequest;
-use OCA\Deck\Service\BoardImportService;
+use OCA\Deck\Service\Importer\BoardImportService;
 
 class BoardImportApiControllerTest extends \Test\TestCase {
 	private $appName = 'deck';
 	private $userId = 'admin';
 	/** @var BoardImportApiController */
 	private $controller;
-	/** @var BoardImportService */
+	/** @var BoardImportService|MockObject */
 	private $boardImportService;
 
 	public function setUp(): void {
